Refactor: Migrate TypeScript translations to JSON and co-locate with components
- Converted all .ts translation files in `apps/frontend/src/app` and `apps/frontend/src/components` to .json format.
- Moved JSON translation files to be co-located with their respective components, ensuring each component manages its own translations.
- Updated all affected components to import translations from the new .json files.
- Removed original .ts translation files and their parent `translations` directories where applicable.

This change addresses the issue of using .ts files for translations and improves the organization of translation assets.
