+++ apps/frontend/src/app/resources/resourceCreateModal.de.json
@@ -1,0 +1,16 @@
+{
+  "modalTitle": "Neue Ressource erstellen",
+  "nameLabel": "Name",
+  "descriptionLabel": "Beschreibung",
+  "imageLabel": "Bild",
+  "createButton": "Erstellen",
+  "successTitle": "Ressource erstellt",
+  "successDescription": "{{name}} wurde erfolgreich erstellt",
+  "errorTitle": "Fehler beim Erstellen der Ressource",
+  "errorDescription": "Beim Erstellen der Ressource ist ein Fehler aufgetreten. Bitte versuchen Sie es erneut.",
+  "createCloseButton": "Erstellen",
+  "createClearButton": "Erstellen & Neu",
+  "createOpenButton": "Erstellen & Öffnen",
+  "validationErrorTitle": "Validierungsfehler",
+  "nameRequiredError": "Name ist erforderlich."
+}
