+++ apps/frontend/src/app/resources/resourceCreateModal.en.json
@@ -1,0 +1,16 @@
+{
+  "modalTitle": "Create a new Resource",
+  "nameLabel": "Name",
+  "descriptionLabel": "Description",
+  "imageLabel": "Image",
+  "createButton": "Create",
+  "successTitle": "Resource created",
+  "successDescription": "{{name}} has been successfully created",
+  "errorTitle": "Failed to create resource",
+  "errorDescription": "An error occurred while creating the resource. Please try again.",
+  "createCloseButton": "Create",
+  "createClearButton": "Create and New",
+  "createOpenButton": "Create and Open",
+  "validationErrorTitle": "Validation Error",
+  "nameRequiredError": "Name is required."
+}
