+++ apps/frontend/src/app/resources/resourceList.de.json
@@ -1,0 +1,5 @@
+{
+  "errorLoadingResources": "Fehler beim Laden der Ressourcen",
+  "noResourcesFound": "Keine Ressourcen gefunden",
+  "ungrouped": "Ungruppiert"
+}
