+++ apps/frontend/src/app/resources/resourceList.en.json
@@ -1,0 +1,5 @@
+{
+  "errorLoadingResources": "Error loading resources",
+  "noResourcesFound": "No resources found",
+  "ungrouped": "Ungrouped"
+}
