+++ apps/frontend/src/app/resources/toolbar.de.json
@@ -1,0 +1,5 @@
+{
+  "searchPlaceholder": "Ressourcen suchen...",
+  "addResource": "Ressource hinzufügen",
+  "addGroup": "Gruppe hinzufügen"
+}
