+++ apps/frontend/src/app/resources/toolbar.en.json
@@ -1,0 +1,5 @@
+{
+  "searchPlaceholder": "Search resources...",
+  "addResource": "Add Resource",
+  "addGroup": "Add Group"
+}
