--- apps/frontend/src/app/resources/translations/resourceCreateModal.de.ts
@@ -1,16 +1,0 @@
-export default {
-  modalTitle: 'Neue Ressource erstellen',
-  nameLabel: 'Name',
-  descriptionLabel: 'Beschreibung',
-  imageLabel: 'Bild',
-  createButton: 'Erstellen',
-  successTitle: 'Ressource erstellt',
-  successDescription: '{{name}} wurde erfolgreich erstellt',
-  errorTitle: 'Fehler beim Erstellen der Ressource',
-  errorDescription: 'Beim Erstellen der Ressource ist ein Fehler aufgetreten. Bitte versuchen Sie es erneut.',
-  createCloseButton: 'Erstellen',
-  createClearButton: 'Erstellen & Neu',
-  createOpenButton: 'Erstellen & Öffnen',
-  validationErrorTitle: 'Validierungsfehler',
-  nameRequiredError: 'Name ist erforderlich.',
-};
