--- apps/frontend/src/app/resources/translations/resourceList.de.ts
@@ -1,5 +1,0 @@
-export default {
-  errorLoadingResources: 'Fehler beim Laden der Ressourcen',
-  noResourcesFound: 'Keine Ressourcen gefunden',
-  ungrouped: 'Ungruppiert',
-};
