--- apps/frontend/src/app/resources/translations/resourceList.en.ts
@@ -1,5 +1,0 @@
-export default {
-  errorLoadingResources: 'Error loading resources',
-  noResourcesFound: 'No resources found',
-  ungrouped: 'Ungrouped',
-};
