--- apps/frontend/src/app/resources/translations/toolbar.de.ts
@@ -1,7 +1,0 @@
-const de = {
-  searchPlaceholder: 'Ressourcen suchen...',
-  addResource: 'Ressource hinzufügen',
-  addGroup: 'Gruppe hinzufügen',
-};
-
-export default de;
