--- apps/frontend/src/app/resources/translations/toolbar.en.ts
@@ -1,7 +1,0 @@
-const en = {
-  searchPlaceholder: 'Search resources...',
-  addResource: 'Add Resource',
-  addGroup: 'Add Group',
-};
-
-export default en;
