+++ apps/frontend/src/app/unauthorized/loginForm.en.json
@@ -1,0 +1,15 @@
+{
+  "title": "Welcome back, maker!",
+  "subtitle": "Ready to bring your next project to life? Sign in to book machines and get creating.",
+  "noAccount": "First time here?",
+  "signUpButton": "Get started here",
+  "username": "Username",
+  "password": "Password",
+  "rememberMe": "Keep me signed in",
+  "forgotPassword": "Password slipped your mind?",
+  "signInButton": "Start making",
+  "error": {
+    "title": "Oops! Sign-in hiccup",
+    "generic": "Please check your credentials and try again"
+  }
+}
