+++ apps/frontend/src/app/unauthorized/registrationForm.de.json
@@ -1,0 +1,22 @@
+{
+  "title": "Bereit zum Gestalten?",
+  "subtitle": "Erstelle dein Konto, um Zeit an unseren 3D-Druckern, Lasercuttern und anderen coolen Maschinen zu buchen. Dein nächstes Projekt ist nur wenige Klicks entfernt!",
+  "hasAccount": "Nutzt du schon unsere Maschinen?",
+  "signInButton": "Hier anmelden",
+  "username": "Wähle einen Benutzernamen",
+  "email": "Deine E-Mail-Adresse",
+  "password": "Erstelle dein Passwort",
+  "passwordConfirmation": "Bestätige dein Passwort",
+  "createAccountButton": "Konto erstellen und loslegen!",
+  "creatingAccount": "Dein Konto wird erstellt...",
+  "error": {
+    "title": "Hoppla! Ein kleiner Fehler",
+    "generic": "Bitte überprüfe deine Angaben und versuche es erneut"
+  },
+  "success": {
+    "title": "Konto erfolgreich erstellt!",
+    "message": "Wir haben eine Aktivierungs-E-Mail an {email} gesendet. Bitte überprüfe deinen Posteingang und klicke auf den Aktivierungslink, um deine Registrierung abzuschließen.",
+    "closeButton": "Verstanden"
+  },
+  "passwordConfirmationError": "Die Passwörter stimmen nicht überein"
+}
