+++ apps/frontend/src/app/unauthorized/registrationForm.en.json
@@ -1,0 +1,22 @@
+{
+  "title": "Ready to create?",
+  "subtitle": "Set up your account to book time on our 3D printers, laser cutters, and other cool machines. Your next project is just a few clicks away!",
+  "hasAccount": "Already using our machines?",
+  "signInButton": "Sign in here",
+  "username": "Pick a username",
+  "email": "Your email address",
+  "password": "Create your password",
+  "passwordConfirmation": "Confirm your password",
+  "createAccountButton": "Create account and start making!",
+  "creatingAccount": "Creating your account...",
+  "error": {
+    "title": "Oops! A small hiccup",
+    "generic": "Please check your info and try again"
+  },
+  "success": {
+    "title": "Account Created Successfully!",
+    "message": "We have sent an activation email to {email}. Please check your inbox and click the activation link to complete your registration.",
+    "closeButton": "Got it"
+  },
+  "passwordConfirmationError": "The passwords do not match"
+}
