+++ apps/frontend/src/app/unauthorized/registrationForm.tsx
@@ -5,8 +5,8 @@ import { Button } from '@heroui/button';
 import { Alert } from '@heroui/alert';
 import { Modal, ModalContent, ModalHeader, ModalBody, ModalFooter, useDisclosure } from '@heroui/modal';
 import { useTranslations } from '@attraccess/plugins-frontend-ui';
-import * as en from './translations/register.en';
-import * as de from './translations/register.de';
+import * as en from './registrationForm.en.json';
+import * as de from './registrationForm.de.json';
 import { useUsersServiceCreateOneUser, UseUsersServiceGetAllUsersKeyFn } from '@attraccess/react-query-client';
 import { useQueryClient } from '@tanstack/react-query';
