+++ apps/frontend/src/app/unauthorized/ssoLogin.de.json
@@ -1,0 +1,3 @@
+{
+  "loginWith": "Anmelden mit {{provider}}"
+}
