+++ apps/frontend/src/app/unauthorized/ssoLogin.en.json
@@ -1,0 +1,3 @@
+{
+  "loginWith": "Login with {{provider}}"
+}
