+++ apps/frontend/src/app/unauthorized/ssoLogin.tsx
@@ -1,7 +1,7 @@
 import { useTranslations } from '@attraccess/plugins-frontend-ui';
 import { Button, Divider, Link } from '@heroui/react';
-import * as de from './translations/ssoLoginButon.de';
-import * as en from './translations/ssoLoginButton.en';
+import * as de from './ssoLogin.de.json';
+import * as en from './ssoLogin.en.json';
 import { useCallback, useEffect, useMemo, useState } from 'react';
 import { useLocation } from 'react-router-dom';
 import { useSsoServiceGetAllSsoProviders, SSOProvider } from '@attraccess/react-query-client';
