--- apps/frontend/src/app/unauthorized/translations/login.en.ts
@@ -1,16 +1,0 @@
-export default {
-  title: 'Welcome back, maker!',
-  subtitle:
-    'Ready to bring your next project to life? Sign in to book machines and get creating.',
-  noAccount: 'First time here?',
-  signUpButton: 'Get started here',
-  username: 'Username',
-  password: 'Password',
-  rememberMe: 'Keep me signed in',
-  forgotPassword: 'Password slipped your mind?',
-  signInButton: 'Start making',
-  error: {
-    title: 'Oops! Sign-in hiccup',
-    generic: 'Please check your credentials and try again',
-  },
-};
