--- apps/frontend/src/app/unauthorized/translations/ssoLoginButon.de.ts
@@ -1,3 +1,0 @@
-export default {
-  loginWith: 'Anmelden mit {{provider}}',
-};
