--- apps/frontend/src/app/unauthorized/translations/ssoLoginButton.en.ts
@@ -1,3 +1,0 @@
-export default {
-  loginWith: 'Login with {{provider}}',
-};
