+++ apps/frontend/src/app/users/UserManagementPage.tsx
@@ -4,8 +4,8 @@ import { useTranslations, UserSearch } from '@attraccess/plugins-frontend-ui';
 import { useAuth } from '../../hooks/useAuth';
 import { Navigate } from 'react-router-dom';
 import { UserPermissionForm } from './components/UserPermissionForm';
-import * as en from './translations/en';
-import * as de from './translations/de';
+import * as en from './userManagement.en.json';
+import * as de from './userManagement.de.json';
 import { Card } from '@heroui/react';
 import { Users, Loader } from 'lucide-react';
 import { useUsersServiceGetOneUserById } from '@attraccess/react-query-client';
