--- apps/frontend/src/app/users/translations/de.ts
@@ -1,40 +1,0 @@
-export const title = 'Benutzerverwaltung';
-export const subtitle =
-  'Verwalten Sie Benutzerberechtigungen und Zugriffskontrolle';
-export const permissions = 'Benutzerberechtigungen';
-export const permissionManagement = 'Berechtigungsverwaltung';
-export const canManageResources = 'Kann Ressourcen verwalten';
-export const canManageSystemConfiguration =
-  'Kann Systemkonfiguration verwalten';
-export const canManageUsers = 'Kann Benutzer verwalten';
-export const saveChanges = 'Änderungen speichern';
-export const userPermissionsSaved =
-  'Benutzerberechtigungen erfolgreich gespeichert';
-export const selectUser =
-  'Wählen Sie einen Benutzer zur Verwaltung von Berechtigungen aus';
-export const filterByPermission = 'Nach Berechtigung filtern';
-export const allPermissions = 'Alle Berechtigungen';
-export const managingUser = 'Benutzer verwalten';
-export const search = 'Benutzer suchen';
-export const noPermissionsGranted = 'Keine Berechtigungen erteilt';
-export const errorUpdatingPermissions =
-  'Fehler beim Aktualisieren der Berechtigungen';
-
-export default {
-  title,
-  subtitle,
-  permissions,
-  permissionManagement,
-  canManageResources,
-  canManageSystemConfiguration,
-  canManageUsers,
-  saveChanges,
-  userPermissionsSaved,
-  selectUser,
-  filterByPermission,
-  allPermissions,
-  managingUser,
-  search,
-  noPermissionsGranted,
-  errorUpdatingPermissions,
-};
