--- apps/frontend/src/app/users/translations/en.ts
@@ -1,35 +1,0 @@
-export const title = 'User Management';
-export const subtitle = 'Manage user permissions and access control';
-export const permissions = 'User Permissions';
-export const permissionManagement = 'Permission Management';
-export const canManageResources = 'Can Manage Resources';
-export const canManageSystemConfiguration = 'Can Manage System Configuration';
-export const canManageUsers = 'Can Manage Users';
-export const saveChanges = 'Save Changes';
-export const userPermissionsSaved = 'User permissions saved successfully';
-export const selectUser = 'Select a user to manage permissions';
-export const filterByPermission = 'Filter by permission';
-export const allPermissions = 'All permissions';
-export const managingUser = 'Managing user';
-export const search = 'Search users';
-export const noPermissionsGranted = 'No permissions granted';
-export const errorUpdatingPermissions = 'Error updating permissions';
-
-export default {
-  title,
-  subtitle,
-  permissions,
-  permissionManagement,
-  canManageResources,
-  canManageSystemConfiguration,
-  canManageUsers,
-  saveChanges,
-  userPermissionsSaved,
-  selectUser,
-  filterByPermission,
-  allPermissions,
-  managingUser,
-  search,
-  noPermissionsGranted,
-  errorUpdatingPermissions,
-};
