+++ apps/frontend/src/app/users/userManagement.de.json
@@ -1,0 +1,18 @@
+{
+  "title": "Benutzerverwaltung",
+  "subtitle": "Verwalten Sie Benutzerberechtigungen und Zugriffskontrolle",
+  "permissions": "Benutzerberechtigungen",
+  "permissionManagement": "Berechtigungsverwaltung",
+  "canManageResources": "Kann Ressourcen verwalten",
+  "canManageSystemConfiguration": "Kann Systemkonfiguration verwalten",
+  "canManageUsers": "Kann Benutzer verwalten",
+  "saveChanges": "Änderungen speichern",
+  "userPermissionsSaved": "Benutzerberechtigungen erfolgreich gespeichert",
+  "selectUser": "Wählen Sie einen Benutzer zur Verwaltung von Berechtigungen aus",
+  "filterByPermission": "Nach Berechtigung filtern",
+  "allPermissions": "Alle Berechtigungen",
+  "managingUser": "Benutzer verwalten",
+  "search": "Benutzer suchen",
+  "noPermissionsGranted": "Keine Berechtigungen erteilt",
+  "errorUpdatingPermissions": "Fehler beim Aktualisieren der Berechtigungen"
+}
