+++ apps/frontend/src/app/users/userManagement.en.json
@@ -1,0 +1,18 @@
+{
+  "title": "User Management",
+  "subtitle": "Manage user permissions and access control",
+  "permissions": "User Permissions",
+  "permissionManagement": "Permission Management",
+  "canManageResources": "Can Manage Resources",
+  "canManageSystemConfiguration": "Can Manage System Configuration",
+  "canManageUsers": "Can Manage Users",
+  "saveChanges": "Save Changes",
+  "userPermissionsSaved": "User permissions saved successfully",
+  "selectUser": "Select a user to manage permissions",
+  "filterByPermission": "Filter by permission",
+  "allPermissions": "All permissions",
+  "managingUser": "Managing user",
+  "search": "Search users",
+  "noPermissionsGranted": "No permissions granted",
+  "errorUpdatingPermissions": "Error updating permissions"
+}
