+++ apps/frontend/src/components/deleteConfirmation.en.json
@@ -1,0 +1,6 @@
+{
+  "title": "Confirm Deletion",
+  "message": "Are you sure you want to delete <bold>{{itemName}}</bold>?<alert>This action cannot be undone.</alert>",
+  "cancelButton": "Cancel",
+  "deleteButton": "Delete"
+}
