--- apps/frontend/src/components/translations/deleteConfirmation.de.ts
@@ -1,7 +1,0 @@
-export default {
-  title: 'Löschen bestätigen',
-  message:
-    'Sind Sie sicher, dass Sie <bold>{{itemName}}</bold> löschen möchten?<alert>Diese Aktion kann nicht rückgängig gemacht werden.</alert>',
-  cancelButton: 'Abbrechen',
-  deleteButton: 'Löschen',
-};
