--- apps/frontend/src/components/translations/deleteConfirmation.en.ts
@@ -1,7 +1,0 @@
-export default {
-  title: 'Confirm Deletion',
-  message:
-    'Are you sure you want to delete <bold>{{itemName}}</bold>?<alert>This action cannot be undone.</alert>',
-  cancelButton: 'Cancel',
-  deleteButton: 'Delete',
-};
