--- apps/frontend/src/components/translations/fileUpload.en.ts
@@ -1,11 +1,0 @@
-export default {
-  label: 'File upload',
-  dragAndDrop: 'Drag and drop an image here, or click to select',
-  acceptedFormats: 'Accepted formats: JPEG, PNG, WebP (max {{maxSize}}MB)',
-  preview: '{{fileName}} ({{fileSize}}MB)',
-  invalidFileType: 'Invalid file type',
-  invalidFileTypeDescription:
-    'Please upload a valid image file ({{allowedTypes}})',
-  fileTooLarge: 'File too large',
-  fileTooLargeDescription: 'File size must be less than {{maxSize}}MB',
-};
